Disable adios2-config script
The reason is that it is being flaky, this has been observed in the
errors for the Install.Make.* tests. We must disable this script and its
tests since otherwise users might expect this to work properly.

This means that Makefile users will not be able to use adios2-config to
get ADIOS2 C,CXX,Fortran building flags. This should not be a big issue
since we expect our users to use CMake.